requirements.txt
APScheduler==3.9.1
backports.zoneinfo
cassandra-driver==3.25.0
certifi==2021.10.8
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
geomet==0.2.1.post1
gunicorn==20.1.0
importlib-metadata==4.11.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.21.5
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
Werkzeug==2.0.3
wincertstore==0.2
zipp==3.7.0
